# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 倾境壁纸 API
  version: 2.4.0
  license:
    name: Proprietary
    url: https://github.com/Jerry1254/wallpapersapp
  description: |
    倾境壁纸正式 App 与管理后台共用的 V1 契约。

    所有 Long ID 在 JSON 中使用字符串；所有时间均为带时区的 ISO 8601 字符串。
    公开目录仅返回封面和能力描述，不返回存储键、绝对路径、内容密钥或正式资源永久地址。
servers:
  - url: /api/v1
    description: 当前来源下的 V1 API；本地联调默认由开发代理转发到 API Server
security: []
...
paths:
...
  /admin/wallpapers:
    get:
      tags: [Admin Catalog]
      operationId: listAdminWallpapers
      summary: 查询管理端壁纸列表
      security:
        - adminCookie: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WallpaperStatus'
        - name: accessType
          in: query
          schema:
            $ref: '#/components/schemas/WallpaperAccessType'
        - name: categoryId
          in: query
          schema:
            $ref: '#/components/schemas/LongId'
        - name: q
          in: query
          schema:
            type: string
            maxLength: 40
      responses:
        '200':
          description: 管理壁纸分页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWallpaperPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Admin Catalog]
      operationId: createAdminWallpaper
      summary: 创建壁纸草稿
      security:
        - adminCookie: []
          adminCsrf: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WallpaperWriteRequest'
      responses:
        '201':
          description: 壁纸草稿已创建
          headers:
            ETag:
              $ref: '#/components/headers/EntityTag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWallpaperDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/wallpapers/{wallpaperId}:
    get:
      tags: [Admin Catalog]
      operationId: getAdminWallpaper
      summary: 查询壁纸、平台变体与资源版本
      security:
        - adminCookie: []
      parameters:
        - $ref: '#/components/parameters/WallpaperId'
      responses:
        '200':
          description: 管理壁纸详情
          headers:
            ETag:
              $ref: '#/components/headers/EntityTag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWallpaperDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Admin Catalog]
      operationId: updateAdminWallpaper
      summary: 修改可编辑的壁纸作品字段
      security:
        - adminCookie: []
          adminCsrf: []
      parameters:
        - $ref: '#/components/parameters/WallpaperId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WallpaperWriteRequest'
      responses:
        '200':
          description: 壁纸已更新
          headers:
            ETag:
              $ref: '#/components/headers/EntityTag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWallpaperDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Admin Catalog]
      operationId: deleteAdminWallpaperDraft
      summary: 硬删除没有任何引用的壁纸草稿
      security:
        - adminCookie: []
          adminCsrf: []
      parameters:
        - $ref: '#/components/parameters/WallpaperId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '204':
          description: 草稿已删除
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/wallpapers/{wallpaperId}/publish:
    post:
      tags: [Admin Catalog]
      operationId: publishAdminWallpaper
      summary: 发布或重新发布壁纸及指定资源版本
      security:
        - adminCookie: []
          adminCsrf: []
      parameters:
        - $ref: '#/components/parameters/WallpaperId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishWallpaperRequest'
      responses:
        '200':
          description: 壁纸已发布，指定版本变为不可变 PUBLISHED
          headers:
            ETag:
              $ref: '#/components/headers/EntityTag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWallpaperDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/wallpapers/{wallpaperId}/offline:
    post:
      tags: [Admin Catalog]
      operationId: offlineAdminWallpaper
      summary: 下架壁纸，停止公开展示和新增兑换
      security:
        - adminCookie: []
          adminCsrf: []
      parameters:
        - $ref: '#/components/parameters/WallpaperId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateChangeReasonRequest'
      responses:
        '200':
          description: 壁纸已下架；已有权益保留，公开目录和新增兑换关闭，App 新下载及旧票据读取停止；已安装资源可继续使用。
          headers:
            ETag:
              $ref: '#/components/headers/EntityTag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWallpaperDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/wallpapers/{wallpaperId}/archive:
    post:
      tags: [Admin Catalog]
      operationId: archiveAdminWallpaper
      summary: 归档草稿或已下架壁纸
      security:
        - adminCookie: []
          adminCsrf: []
      parameters:
        - $ref: '#/components/parameters/WallpaperId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateChangeReasonRequest'
      responses:
        '200':
          description: 壁纸已归档
          headers:
            ETag:
              $ref: '#/components/headers/EntityTag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWallpaperDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/wallpapers/{wallpaperId}/variants:
    post:
      tags: [Admin Catalog]
      operationId: createAdminWallpaperVariant
      summary: 为壁纸创建平台变体
      security:
        - adminCookie: []
          adminCsrf: []
      parameters:
        - $ref: '#/components/parameters/WallpaperId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VariantWriteRequest'
      responses:
        '201':
          description: 平台变体已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWallpaperVariant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/variants/{variantId}:
    patch:
      tags: [Admin Catalog]
      operationId: updateAdminWallpaperVariant
      summary: 修改平台变体约束
      security:
        - adminCookie: []
          adminCsrf: []
      parameters:
        - $ref: '#/components/parameters/VariantId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VariantWriteRequest'
      responses:
        '200':
          description: 变体已更新
          headers:
            ETag:
              $ref: '#/components/headers/EntityTag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWallpaperVariant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Admin Catalog]
      operationId: deleteAdminWallpaperVariant
      summary: 删除从未发布且没有历史引用的平台变体
      security:
        - adminCookie: []
          adminCsrf: []
      parameters:
        - $ref: '#/components/parameters/VariantId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '204':
          description: 变体已删除
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/variants/{variantId}/resource-versions:
    post:
      tags: [Admin Catalog]
      operationId: createAdminResourceVersion
      summary: 使用已上传 Asset 创建不可变候选资源版本
      security:
        - adminCookie: []
          adminCsrf: []
      parameters:
        - $ref: '#/components/parameters/VariantId'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceVersionRequest'
      responses:
        '201':
          description: 候选资源版本已创建并进入校验
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResourceVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/resource-versions/{resourceVersionId}/secure-package:
    post:
      tags: [Admin Catalog]
      operationId: buildAdminSecurePackage
      summary: 为 Android 兼容资源版本制作正式包与受限试用包
      description: 校验媒体后为 READY 版本事务内制作 format 2 正式包和 format 3 APP_PREVIEW 受限包，失败回滚并清理本次对象。两包都存在时幂等返回；已有正式包的 PUBLISHED 版本允许补建受限包，正式包和正式 manifest 不变。已发布且无正式包的版本仍需创建下一版本。缺少资源签名或媒体工具配置时不可制作。
      security:
        - adminCookie: []
          adminCsrf: []
      parameters:
        - $ref: '#/components/parameters/ResourceVersionId'
      responses:
        '200':
          description: 已制作版本与真实 manifest 摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResourceVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/resource-versions/{resourceVersionId}:
    get:
      tags: [Admin Catalog]
      operationId: getAdminResourceVersion
      summary: 查询资源版本、绑定与校验结果
      security:
        - adminCookie: []
      parameters:
        - $ref: '#/components/parameters/ResourceVersionId'
      responses:
        '200':
          description: 资源版本详情
          headers:
            ETag:
              $ref: '#/components/headers/EntityTag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResourceVersion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    deviceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-device-session
      description: Redis 中可撤销、可过期的短期设备会话。
    downloadTicketBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-download-ticket
      description: 只允许读取一个资源版本的短时票据；放在 Authorization header，避免进入 URL 日志。
    adminCookie:
      type: apiKey
      in: cookie
      name: QJ_ADMIN_SESSION
      description: HttpOnly、Secure、SameSite Cookie。
    adminCsrf:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 管理端写操作必须与会话绑定的 CSRF token 一致。
    adminDeliveryTicket:
      type: apiKey
      in: header
      name: X-Delivery-Ticket
      description: 只允许读取一次性交付 CSV 的短时票据，确认交付或过期后失效。
    previewTicketBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-preview-ticket
      description: 独立 90 秒预览票据，仅允许读取受限预览包；放在 Authorization header，禁止 URL 携带。
  headers:
    EntityTag:
      description: 当前聚合 lockVersion 的强 ETag，例如 "7"。
      schema:
        type: string
        pattern: '^"[0-9]+"$'
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    WallpaperId:
      name: wallpaperId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/LongId'
...
    VariantId:
      name: variantId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/LongId'
    ResourceVersionId:
      name: resourceVersionId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/LongId'
...
    RequestId:
      name: X-Request-Id
      in: header
      description: 可选链路请求 UUID；缺省时由服务端生成并在错误响应返回。
      schema:
        type: string
        format: uuid
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 上次读取到的强 ETag，用于阻止覆盖并发修改。
      schema:
        type: string
        pattern: '^"[0-9]+"$'
...
  responses:
    BadRequest:
      description: 请求格式、参数或字段校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: 会话或身份凭据缺失、无效或已过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: 已识别身份无权执行操作，或请求证明/CSRF 校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: 资源不存在或当前调用方不可见
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: 唯一键、状态、引用关系或幂等键与请求内容冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    PreconditionFailed:
      description: If-Match 与当前 lockVersion 不一致
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    UnprocessableEntity:
      description: 请求格式正确但不满足领域或安全校验规则
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    TooManyRequests:
      description: 超过设备、IP、账号、码摘要或下载频率限制
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: 未预期服务错误；不返回堆栈、秘密或内部路径
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    LongId:
      type: string
      pattern: '^[1-9][0-9]*$'
      example: '42'
      description: Java Long ID 的 JSON 字符串表示。
...
    WallpaperAccessType:
      type: string
      description: 正式资源获取方式。FREE 可直接申请正式下载；REDEEM 必须先取得设备权益。
      enum: [REDEEM, FREE]
...
    WallpaperStatus:
      type: string
      enum: [DRAFT, PUBLISHED, OFFLINE, ARCHIVED]
    DeliveryPlatform:
      type: string
      enum: [ANDROID, IOS, HARMONYOS, UNIVERSAL]
...
    ResourceType:
      type: string
      enum: [LAYER_PARALLAX, VIDEO, LIVE_PHOTO, STATIC_IMAGE, THEME_PACKAGE]
...
    WallpaperWriteRequest:
      type: object
      required:
        [
          title,
          slug,
          accessType,
          rootCategoryId,
          coverAssetId,
          sortOrder,
          copyrightNote
        ]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 40
        slug:
          type: string
          minLength: 2
          maxLength: 64
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        accessType:
          $ref: '#/components/schemas/WallpaperAccessType'
        rootCategoryId:
          $ref: '#/components/schemas/LongId'
        childCategoryId:
          type: string
          allOf:
            - $ref: '#/components/schemas/LongId'
          nullable: true
          description: 提供时必须是 rootCategoryId 的直接子节点；数据库只保存这个最深节点。
        coverAssetId:
          $ref: '#/components/schemas/LongId'
        featuredRank:
          type: integer
          minimum: 0
          maximum: 999999
          nullable: true
        sortOrder:
          type: integer
          minimum: 0
          maximum: 999999
        copyrightNote:
          type: string
          minLength: 1
          maxLength: 500
...
    AdminWallpaperDetail:
      allOf:
        - $ref: '#/components/schemas/AdminWallpaperSummary'
        - type: object
          required: [copyrightNote, variants]
          properties:
            copyrightNote:
              type: string
              maxLength: 500
            variants:
              type: array
              items:
                $ref: '#/components/schemas/AdminWallpaperVariant'
    AdminWallpaperPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminWallpaperSummary'
        page:
          $ref: '#/components/schemas/PageMetadata'
    VariantWriteRequest:
      type: object
      required: [platform, resourceType, enabled]
      properties:
        platform:
          $ref: '#/components/schemas/DeliveryPlatform'
        resourceType:
          $ref: '#/components/schemas/ResourceType'
        minimumOsVersion:
          type: string
          maxLength: 32
          nullable: true
        capabilityRequirements:
          type: array
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            maxLength: 64
        enabled:
          type: boolean
          description: 管理端是否将该变体纳入下一次发布；已发布内容只在执行发布动作时切换。
    AdminWallpaperVariant:
      type: object
      required:
        [
          id,
          platform,
          resourceType,
          capabilityRequirements,
          enabled,
          resourceVersions,
          version
        ]
      properties:
        id:
          $ref: '#/components/schemas/LongId'
        platform:
          $ref: '#/components/schemas/DeliveryPlatform'
        resourceType:
          $ref: '#/components/schemas/ResourceType'
        minimumOsVersion:
          type: string
          maxLength: 32
          nullable: true
        capabilityRequirements:
          type: array
          items:
            type: string
            maxLength: 64
        enabled:
          type: boolean
        resourceVersions:
          type: array
          items:
            $ref: '#/components/schemas/AdminResourceVersion'
        version:
          type: integer
          format: int64
          minimum: 0
...
    CreateResourceVersionRequest:
      type: object
      required: [versionNo, bindings]
      properties:
        versionNo:
          type: integer
          minimum: 1
          maximum: 2147483647
        manifestSha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
          nullable: true
        bindings:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/CreateResourceBindingRequest'
...
    AdminResourceVersion:
      type: object
      required: [id, versionNo, status, bindings, sourcePackage, createdAt, version]
      properties:
        id:
          $ref: '#/components/schemas/LongId'
        versionNo:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/ResourceVersionStatus'
        manifestSha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
          nullable: true
        validationErrors:
          type: array
          maxItems: 100
          items:
            type: object
            required: [code, message]
            properties:
              code:
                type: string
                maxLength: 64
              message:
                type: string
                maxLength: 300
        bindings:
          type: array
          items:
            $ref: '#/components/schemas/AdminResourceBinding'
        sourcePackage:
          type: object
          allOf:
            - $ref: '#/components/schemas/AdminParallaxSourcePackage'
          nullable: true
          description: ZIP 导入来源；旧版手工创建资源显式返回 null。
        publishedAt:
          type: string
          allOf:
            - $ref: '#/components/schemas/Timestamp'
          nullable: true
        retiredAt:
          type: string
          allOf:
            - $ref: '#/components/schemas/Timestamp'
          nullable: true
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        version:
          type: integer
          format: int64
          minimum: 0
...
    PublishWallpaperRequest:
      type: object
      required: [resourceVersionIds]
      properties:
        resourceVersionIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/LongId'
    StateChangeReasonRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 300
...